+++ opps/containers/forms.py
@@ -8,7 +8,7 @@ from opps.db.models.fields.jsonf import JSONFormField
 from opps.fields.widgets import JSONField
 from opps.fields.models import Field, FieldOption
 
-from .models import Container, ContainerBoxContainers
+from .models import Container, ContainerBoxContainers, ContainerImage
 
 
 class ContainerBoxContainersInlineForm(forms.ModelForm):
@@ -19,6 +19,16 @@ class ContainerBoxContainersInlineForm(forms.ModelForm):
 
     class Meta:
         model = ContainerBoxContainers
+
+
+class ContainerImageInlineForm(forms.ModelForm):
+    def __init__(self, *args, **kwargs):
+        super(ContainerImageInlineForm,
+              self).__init__(*args, **kwargs)
+        self.fields['order'].widget.attrs['readonly'] = True
+
+    class Meta:
+        model = ContainerImage
 
 
 class ContainerAdminForm(forms.ModelForm):
@@ -37,25 +47,25 @@ class ContainerAdminForm(forms.ModelForm):
             if field.type == 'checkbox':
                 for fo in FieldOption.objects.filter(field=field):
                     self.fields[
-                        'json_{}_{}'.format(
+                        'json_{0}_{1}'.format(
                             field.slug, fo.option.slug
                         )] = forms.CharField(required=False)
             else:
                 self.fields[
-                    'json_{}'.format(field.slug)
+                    'json_{0}'.format(field.slug)
                 ] = forms.CharField(required=False)
 
     def clean(self):
-        msg = _('The slug "%s" already exists on channel "%s" at site "%s"')
+        msg = _('The slug "%(slug)s" already exists on channel "%(channel)s"\
+         at site "%(site)s"')
         data = self.cleaned_data
-        if data['site'] and data['channel'] and data['slug']:
+        if data.get('site') and data.get('channel') and data.get('slug'):
             qs = Container.objects.filter(
                 site=data['site'], channel=data['channel'], slug=data['slug'])
             if self.instance.pk is not None:
                 qs = qs.exclude(pk=self.instance.pk)
             if qs.exists():
-                args = (data['slug'], data['channel'], data['site'])
-                msg = msg % args
+                msg = msg % data
                 self._errors["slug"] = self.error_class([msg])
         return data